MIDI improvements
* Seperated Midi objects into their own file.
* Fixed bug in midi._parse_variable_length_quantity()
* Added tests
